+++ fediseer/apis/v1/hesitations.py
@@ -6,6 +6,7 @@ from fediseer import enums
 
 class HesitationsGiven(Resource):
     get_parser = reqparse.RequestParser()
+    get_parser.add_argument("apikey", type=str, required=False, help="An instance's API key.", location='headers')
     get_parser.add_argument("Client-Agent", default="unknown:0:unknown", type=str, required=False, help="The client name and version.", location="headers")
     get_parser.add_argument("csv", required=False, type=bool, help="Set to true to return just the domains as a csv. Mutually exclusive with domains", location="args")
     get_parser.add_argument("domains", required=False, type=bool, help="Set to true to return just the domains as a list. Mutually exclusive with csv", location="args")
@@ -16,17 +17,40 @@ class HesitationsGiven(Resource):
     @cache.cached(timeout=10, query_string=True)
     @api.marshal_with(models.response_model_model_Hesitations_get, code=200, description='Instances', skip_none=True)
     @api.response(404, 'Instance not registered', models.response_model_error)
+    @api.response(401, 'API key not found', models.response_model_error)
+    @api.response(403, 'Access Denied', models.response_model_error)
     def get(self, domains_csv):
         '''Display all hesitations given out by one or more domains
         You can pass a comma-separated list of domain names
         and the results will be a set of all their hesitations together.
         '''
         self.args = self.get_parser.parse_args()
+        get_instance = None
+        if self.args.apikey:
+            get_instance = database.find_instance_by_api_key(self.args.apikey)
+            if not get_instance:
+                raise e.Unauthorized(f"No Instance found matching provided API key. Please ensure you've typed it correctly")
         domains_list = domains_csv.split(',')
-        instances = database.find_multiple_instance_by_domains(domains_list)
-        if not instances:
+        precheck_instances = database.find_multiple_instance_by_domains(domains_list)
+        if not precheck_instances:
             raise e.NotFound(f"No Instances found matching any of the provided domains. Have you remembered to register them?")
-        if self.args.min_hesitations > len(domains_list):
+        instances = []
+        for instance in precheck_instances:
+            if instance.visibility_hesitations == enums.ListVisibility.ENDORSED:
+                if get_instance is None:
+                    continue
+                if instance != get_instance and not instance.is_endorsing(get_instance):
+                    continue
+            if instance.visibility_hesitations == enums.ListVisibility.PRIVATE:
+                logger.debug([instance.visibility_hesitations,instance,get_instance,instance != get_instance])
+                if get_instance is None:
+                    continue
+                if instance != get_instance:
+                    continue
+            instances.append(instance)
+        if len(instances) == 0:
+            raise e.Forbidden(f"You do not have access to see these hesitations")
+        if self.args.min_hesitations > len(instances):
             raise e.BadRequest(f"You cannot request more hesitations than the amount of reference domains")
         instance_details = []
         for c_instance in database.get_all_dubious_instances_by_hesitant_id([instance.id for instance in instances]):
@@ -63,6 +87,7 @@ class HesitationsGiven(Resource):
 
 class Hesitations(Resource):
     get_parser = reqparse.RequestParser()
+    get_parser.add_argument("apikey", type=str, required=False, help="An instance's API key.", location='headers')
     get_parser.add_argument("Client-Agent", default="unknown:0:unknown", type=str, required=False, help="The client name and version.", location="headers")
     get_parser.add_argument("csv", required=False, type=bool, help="Set to true to return just the domains as a csv. Mutually exclusive with domains", location="args")
     get_parser.add_argument("domains", required=False, type=bool, help="Set to true to return just the domains as a list. Mutually exclusive with csv", location="args")
@@ -75,11 +100,30 @@ class Hesitations(Resource):
         '''Display all hesitations received by a specific domain
         '''
         self.args = self.get_parser.parse_args()
+        get_instance = None
+        if self.args.apikey:
+            get_instance = database.find_instance_by_api_key(self.args.apikey)
+            if not get_instance:
+                raise e.Unauthorized(f"No Instance found matching provided API key. Please ensure you've typed it correctly")
         instance = database.find_instance_by_domain(domain)
         if not instance:
             raise e.NotFound(f"No Instance found matching provided domain. Have you remembered to register it?")
+        precheck_instances = database.get_all_hesitant_instances_by_dubious_id(instance.id)
+        instances = []
+        for instance in precheck_instances:
+            if instance.visibility_endorsements == enums.ListVisibility.ENDORSED:
+                if get_instance is None:
+                    continue
+                if not instance.is_endorsing(get_instance):
+                    continue
+            if instance.visibility_endorsements == enums.ListVisibility.PRIVATE:
+                if get_instance is None:
+                    continue
+                if not instance != get_instance:
+                    continue
+            instances.append(instance)
         instance_details = []
-        for c_instance in database.get_all_hesitant_instances_by_dubious_id(instance.id):
+        for c_instance in instances:
             hesitations = database.get_all_hesitation_reasons_for_dubious_id(instance.id, [c_instance.id])
             hesitations = [c for c in hesitations if c.reason is not None]
             c_instance_details = c_instance.get_details()
@@ -149,9 +193,12 @@ class Hesitations(Resource):
             evidence=evidence,
         )
         db.session.add(new_hesitation)
+        target_domain = target_instance.domain
+        if instance.visibility_hesitations != enums.ListVisibility.OPEN:
+            target_domain = '[REDACTED]'
         new_report = Report(
             source_domain=instance.domain,
-            target_domain=target_instance.domain,
+            target_domain=target_domain,
             report_type=enums.ReportType.HESITATION,
             report_activity=enums.ReportActivity.ADDED,
         )
@@ -206,9 +253,12 @@ class Hesitations(Resource):
                 changed = True
         if changed is False:
             return {"message":'OK'}, 200
+        target_domain = target_instance.domain
+        if instance.visibility_hesitations != enums.ListVisibility.OPEN:
+            target_domain = '[REDACTED]'
         new_report = Report(
             source_domain=instance.domain,
-            target_domain=target_instance.domain,
+            target_domain=target_domain,
             report_type=enums.ReportType.HESITATION,
             report_activity=enums.ReportActivity.MODIFIED,
         )
@@ -245,9 +295,12 @@ class Hesitations(Resource):
         if not hesitation:
             return {"message":'OK'}, 200
         db.session.delete(hesitation)
+        target_domain = target_instance.domain
+        if instance.visibility_hesitations != enums.ListVisibility.OPEN:
+            target_domain = '[REDACTED]'
         new_report = Report(
             source_domain=instance.domain,
-            target_domain=target_instance.domain,
+            target_domain=target_domain,
             report_type=enums.ReportType.HESITATION,
             report_activity=enums.ReportActivity.DELETED,
         )